fix(telegram): map presentation slash buttons to native callbacks
