+++ extensions/telegram/src/command-ui.ts
@@ -5,6 +5,7 @@ import {
   buildProviderKeyboard,
   type ProviderInfo,
 } from "./model-buttons.js";
+import { buildTelegramNativeCommandCallbackData } from "./native-command-callback-data.js";
 
 export function buildCommandsPaginationKeyboard(
   currentPage: number,
@@ -94,7 +95,7 @@ export function buildTelegramModelsAddProviderChannelData(params: {
   const buttons = params.providers.map((provider) => [
     {
       text: provider.id,
-      callback_data: `/models add ${provider.id}`,
+      callback_data: buildTelegramNativeCommandCallbackData(`/models add ${provider.id}`),
     },
   ]);
   return {
